Unindo duas activities redundantes (ReportActivity e FoodReportActivity), adicionando funcionalidade de pesquisa à tabela feedings (Alimentação) no banco e completando código de gerenciamento de abas para variar entre sono e alimentação
