--- app/src/main/res/layout/activity_food_report.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main2"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/bg_color"
-    tools:context=".ReportActivity">
-
-    <ImageView
-        android:id="@+id/imageView0"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="-22dp"
-        android:src="@drawable/bush"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="-22dp"
-        android:rotation="180"
-        android:src="@drawable/bush"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
-
-    <ImageView
-        android:id="@+id/foodContainer"
-        android:layout_width="223dp"
-        android:layout_height="57dp"
-        android:layout_marginBottom="-1dp"
-        android:scaleType="fitXY"
-        app:layout_constraintBottom_toTopOf="@+id/foodScroll"
-        app:layout_constraintStart_toStartOf="@+id/foodScroll"
-        app:srcCompat="@drawable/relatory_title" />
-
-    <TextView
-        android:id="@+id/relatoryTitle"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:fontFamily="@font/leaguespartan_regular"
-        android:gravity="center|center_vertical"
-        android:text="Histórico de Refeições"
-        android:textAlignment="center"
-        android:textColor="@color/black"
-        android:textSize="20sp"
-        android:typeface="normal"
-        app:layout_constraintBottom_toBottomOf="@+id/foodContainer"
-        app:layout_constraintEnd_toEndOf="@+id/foodContainer"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="@+id/foodContainer"
-        app:layout_constraintTop_toTopOf="@+id/foodContainer"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <ScrollView
-        android:id="@+id/foodScroll"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:background="@drawable/relatory_bg"
-        android:visibility="visible"
-        android:padding="14dp"
-        android:layout_marginLeft="24dp"
-        android:layout_marginRight="24dp"
-        android:layout_marginTop="60dp"
-        app:layout_constraintBottom_toTopOf="@+id/imageView2"
-        app:layout_constraintEnd_toEndOf="@+id/imageView0"
-        app:layout_constraintStart_toStartOf="@+id/imageView0"
-        app:layout_constraintTop_toBottomOf="@+id/imageView0">
-
-        <LinearLayout
-        android:id="@+id/foodItemsContainer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-        </LinearLayout>
-    </ScrollView>
-
-    <ImageView
-        android:id="@+id/imageView5"
-        android:layout_width="128dp"
-        android:layout_height="57dp"
-        android:scaleType="fitEnd"
-        app:layout_constraintBottom_toTopOf="@+id/foodScroll"
-        app:layout_constraintEnd_toEndOf="@+id/foodScroll"
-        app:layout_constraintTop_toTopOf="@+id/foodButton"
-        app:layout_constraintVertical_bias="0.0"
-        app:srcCompat="@drawable/options_bg" />
-
-    <ImageButton
-        android:id="@+id/foodButton"
-        android:layout_width="64dp"
-        android:layout_height="57dp"
-        android:backgroundTint="#00FFFFFF"
-        android:scaleType="center"
-        app:layout_constraintStart_toStartOf="@+id/imageView5"
-        app:layout_constraintBottom_toTopOf="@+id/foodScroll"
-        app:layout_constraintVertical_bias="1.0"
-        app:srcCompat="@drawable/relatory_selected" />
-
-    <ImageView
-        android:id="@+id/foodIcon"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:scaleType="centerInside"
-        app:layout_constraintEnd_toEndOf="@+id/foodButton"
-        app:layout_constraintStart_toStartOf="@+id/foodButton"
-        app:layout_constraintBottom_toTopOf="@+id/foodScroll"
-        app:layout_constraintTop_toTopOf="@id/imageView5"
-        app:srcCompat="@drawable/food_draw" />
-
-    <ImageButton
-        android:id="@+id/sleepButton"
-        android:layout_width="64dp"
-        android:layout_height="57dp"
-        android:backgroundTint="#00FFFFFF"
-        android:scaleType="center"
-        app:layout_constraintBottom_toBottomOf="@+id/imageView5"
-        app:layout_constraintEnd_toEndOf="@+id/imageView5"
-        app:layout_constraintTop_toTopOf="@+id/imageView5"
-        app:srcCompat="@drawable/relatory_unselected" />
-
-    <ImageView
-        android:id="@+id/sleepIcon"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:scaleType="centerInside"
-        app:layout_constraintBottom_toBottomOf="@+id/foodButton"
-        app:layout_constraintEnd_toEndOf="@+id/imageView5"
-        app:layout_constraintStart_toStartOf="@+id/sleepButton"
-        app:layout_constraintTop_toTopOf="@+id/foodButton"
-        app:srcCompat="@drawable/sleep_draw" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
